+++ CHANGELOG.md
@@ -6,7 +6,6 @@
 * Fixed #PAC-280: Extension library requires `etc` configuration directory
 * Fixed #PAC-206: Prevent finder mappings of different libraries to be overwritten
 * Fixed #PAC-318: Serialization exception when using a custom database configuration
-* Add #PAC-486: Add Getter and Setter for `--log-file` commandline parameter 
 
 ## Features
 
@@ -14,6 +13,7 @@
 * Add #PAC-57: Deleting dedicated attribute values via import by setting a configurable value
 * Add #PAC-349: Support for JMS serializer v2.0 + v3.0
 * Configuration for import explizit filename
+* Add #PAC-294: integration of the strict mode
 
 # Version 16.1.0